requirements.txt
Pillow
nltk
tqdm
block.bootstrap.pytorch
murel.bootstrap.pytorch
konlpy
mecab-python3
gensim
